@@ -4,7 +4,9 @@ description: Flutter package to render html as widgets that focuses on correctne
 homepage: https://github.com/daohoangson/flutter_widget_from_html/tree/master/packages/core
 
 environment:
-  flutter: ">=3.3.0"
+  flutter: ">=3.7.0-0"
+
+  # flutter@3.7 uses sdk@2.19 but our test scripts are reused in earlier SDK
   sdk: ">=2.13.0 <3.0.0"
 
 dependencies:
