@@ -5,7 +5,7 @@ homepage: https://github.com/daohoangson/flutter_widget_from_html
 
 environment:
   flutter: ">=2.2.0"
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.13.0 <3.0.0"
 
 dependencies:
   flutter:
